move root search to seperate util file, cache dir moved to noop dir in home dir
