+++ lib/commands/reset.js
@@ -1,51 +1,18 @@
 const discovery = require('noop-discovery')
 const async = require('async')
-const path = require('path')
 const fs = require('fs')
 const chalk = require('chalk')
 const rimraf = require('rimraf')
 const os = require('os')
+const findRoot = require('./util/findRoot')
 
 module.exports = (argv) => {
   if (!argv.resource.length) return console.log(chalk.red('Missing operand \'resource\''))
   const selectedResources = argv.resource
-  const root =
-    typeof argv.rootPath === 'string'
-      ? path.resolve(
-        path.isAbsolute(argv.rootPath) ? '' : process.cwd(),
-        argv.rootPath
-      )
-      : undefined
+  const rootArgv = argv.rootPath
   async.auto({
     rootDir: (done) => {
-      if (root) {
-        fs.access(root, (err) => {
-          if (err) {
-            done(new Error('Root path could not be found!'))
-          } else {
-            done(null, root)
-          }
-        })
-      } else {
-        const gitRootDir = (dir) => (
-          fs.promises.readdir(dir, { withFileTypes: true })
-            .then((files) => {
-              for (const file of files) {
-                if ((file.name === '.git') && file.isDirectory()) {
-                  return dir
-                }
-              }
-              if (dir !== '/') return gitRootDir(path.resolve(dir, '..'))
-            })
-        )
-        gitRootDir(process.cwd()).then((rootDir) => {
-          if (rootDir) {
-            done(null, rootDir)
-          } else {
-            done(new Error('Unable to locate root of git project!'))
-          }
-        })
-      }
+      findRoot(done, rootArgv)
     },
     app: ['rootDir', (results, done) => {
       console.log(`Application root found at '${results.rootDir}'`)
