+++ lib/commands/run.js
@@ -7,6 +7,7 @@ const fs = require('fs')
 const Inspector = require('../inspector')
 const exec = require('child_process').exec
 const readline = require('readline')
+const findRoot = require('./util/findRoot')
 
 module.exports = (argv) => {
   const port = argv.port
@@ -14,16 +15,10 @@ module.exports = (argv) => {
   const verbose = argv.verbose
   let selectedComponents = argv.component
   let selectedResources = argv.resource
-  const root =
-    typeof argv.rootPath === 'string'
-      ? path.resolve(
-        path.isAbsolute(argv.rootPath) ? '' : process.cwd(),
-        argv.rootPath
-      )
-      : undefined
+  const rootArgv = argv.rootPath
   async.auto({
     rootDir: (done) => {
-      findRootDir(done, root)
+      findRoot(done, rootArgv)
     },
     app: ['rootDir', (results, done) => {
       console.log(`Application root found at '${results.rootDir}'`)
@@ -82,37 +77,6 @@ module.exports = (argv) => {
     if (err) return console.log(chalk.red(err.message))
     process.exit()
   })
-}
-
-const findRootDir = (done, root) => {
-  if (root) {
-    fs.access(root, (err) => {
-      if (err) {
-        done(new Error('Root path could not be found!'))
-      } else {
-        done(null, root)
-      }
-    })
-  } else {
-    const gitRootDir = (dir) => (
-      fs.promises.readdir(dir, { withFileTypes: true })
-        .then((files) => {
-          for (const file of files) {
-            if ((file.name === '.git') && file.isDirectory()) {
-              return dir
-            }
-          }
-          if (dir !== '/') return gitRootDir(path.resolve(dir, '..'))
-        })
-    )
-    gitRootDir(process.cwd()).then((rootDir) => {
-      if (!rootDir) {
-        done(new Error('Unable to locate root of git project!'))
-      } else {
-        done(null, rootDir)
-      }
-    })
-  }
 }
 
 const parseEnvOverrides = (statements = [], files = [], rootDir, done) => {
