+++ lib/devServer.js
@@ -11,6 +11,8 @@ const Docker = require('dockerode')
 const request = require('request')
 const fs = require('fs')
 const moment = require('moment')
+const os = require('os')
+const mkdirp = require('mkdirp')
 
 const docker = new Docker({
   socketPath: '/var/run/docker.sock'
@@ -334,35 +336,41 @@ class DevServer {
 
   refreshCertificates (done) {
     const url = 'https://noop-local.s3.amazonaws.com/certificates/localnoop.app/latest.json'
-    const cacheDir = path.resolve(__dirname, '../cache')
+    const cacheDir = path.resolve(os.homedir(), '.noop', 'cache')
     const cacheFile = path.resolve(cacheDir, 'certificate.json')
-    if (!fs.existsSync(cacheDir)) fs.mkdirSync(cacheDir)
-    request({ url, timeout: 5000 }, (err, res, body) => {
-      if (res && res.statusCode > 200) err = new Error(`Invalid response retrieving certificate [${res.statusCode}]`)
+    fs.access(cacheDir, (err) => {
       if (err) {
-        console.log(`Unable to refresh local certificates, using cache. ${err.message}`)
-        fs.readFile(cacheFile, (err, data) => {
-          if (err) return done(new Error('Error preparing local server certificates.'))
+        mkdirp(cacheDir, (err) => {
+          if (err) return done(err)
+        })
+      }
+      request({ url, timeout: 5000 }, (err, res, body) => {
+        if (res && res.statusCode > 200) err = new Error(`Invalid response retrieving certificate [${res.statusCode}]`)
+        if (err) {
+          console.log(`Unable to refresh local certificates, using cache. ${err.message}`)
+          fs.readFile(cacheFile, (err, data) => {
+            if (err) return done(new Error('Error preparing local server certificates.'))
+            try {
+              const json = JSON.parse(data)
+              this.certificateChain = Buffer.from(json.chain, 'base64').toString()
+              this.certificateKey = Buffer.from(json.key, 'base64').toString()
+              done()
+            } catch (err) {
+              done(err)
+            }
+          })
+        } else {
           try {
-            const json = JSON.parse(data)
+            const json = JSON.parse(body)
             this.certificateChain = Buffer.from(json.chain, 'base64').toString()
             this.certificateKey = Buffer.from(json.key, 'base64').toString()
-            done()
+            console.log(`Caching certificate to '${cacheFile}'`)
+            fs.writeFile(cacheFile, JSON.stringify(json), done)
           } catch (err) {
             done(err)
           }
-        })
-      } else {
-        try {
-          const json = JSON.parse(body)
-          this.certificateChain = Buffer.from(json.chain, 'base64').toString()
-          this.certificateKey = Buffer.from(json.key, 'base64').toString()
-          console.log(`Caching certificate to '${cacheFile}'`)
-          fs.writeFile(cacheFile, JSON.stringify(json), done)
-        } catch (err) {
-          done(err)
         }
-      }
+      })
     })
   }
 }
