+++ package.json
@@ -42,7 +42,7 @@
     "mkdirp": "^0.5.1",
     "moment": "^2.29.1",
     "node-cron": "^2.0.3",
-    "noop-discovery": "^1.3.14",
+    "noop-discovery": "^1.3.15",
     "path-to-regexp": "^2.4.0",
     "request": "^2.88.0",
     "rimraf": "^2.6.2",
